requirements.txt
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djoser==2.3.3
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mysqlclient==2.2.7
tzdata==2025.2
whitenoise==6.11.0
